refactor: Replaced PlayerList with HashTable<Player>
Matches known processing behavior closer.
IP-infleunced processing is inherent with this collection type and usage.

There is no "PID" number.
"Having PID" is where you are in the underlying collection.
You're just in a lower bucket or have logged in sooner in your bucket.

"PID" (processing) should be considered separate from "PID" (protocol). "Slot" is a better term for the protocol (not player-aware).